pyproject.toml
[tool.poetry]
name = "event-booker-backend"
version = "0.1.0"
description = ""
authors = ["WiktorWr <wiktor4411@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
ruff = "^0.6.9"
mypy = "^1.11.2"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.3"
pydantic-settings = "^2.5.2"
factory-boy = "^3.3.1"
pytest = "^8.3.3"
uvloop = "^0.20.0"
httptools = "^0.6.1"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.12"
types-passlib = "^1.7.7.20240819"
pytest-cov = "^5.0.0"
types-python-jose = "^3.3.4.20240106"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
